Common Metadata Repository (CMR) (#2)
* Support to NASA's Common Metadata Repository

* Some basic tests

* Importing specific functions from numpy

* Temporary solution for L3m daily chl_a

* Moving some original functions out of main OceanColor

* Simplifying search rule

* Logging CMR

* A note for an alternative for granule_search()

* refactor: Split search with sparced tracks optimze response

Searching along a sparsed sequence could return excessive potential
targets between waypoints. If that is the case, split targets in time
and recursively search. This can have a huge impact reducing the
potential targets for the matchup procedure.

* test: Checking bloom_filter on spaced targets and duplicates

* fix: if single target bloom_filter doesn't need to split

If is a single target it doesn't need to check if it is sparse.

* refactor: Memory on bloo_filter to avoid redundancy