Coarse inrange crop (#8)
* refactor: Coarse pre crop before checking in range

It was already doing that for latitude, but now it also crop the input
Dataset by longitude. Because the meridians converge poleward, it
required an extra step. The limit is not exact, but intended to cut off
the points clearly off range.

* test: Testing inrange_L2 near longitude 180

* test: Fixing test check since there is a new field: product_name

* feat: Adding product_name into L2 results

This is useful to help to trace back the results.

* fix: No coarse solution if nearby the international day line

If the coarse searching window crosses the day line, don't simplify but
do the (expensive) precise calculations.

* test: inrange simple check using number of matchups

Before it was using size which includes the number of columns. I need to
improve these tests but for now let's not fail if I change the number of
columns (fields) in the output result from searches.

* test: Expanding cached items for CI

* fix: Typo